+++ common/src/main/java/com/daqem/jobsplus/integration/arc/data/type/JobsPlusActionDataType.java
@@ -8,7 +8,7 @@ import net.minecraft.resources.ResourceLocation;
 
 public interface JobsPlusActionDataType<T> extends ActionDataType<T> {
 
-    IActionDataType<Integer> JOB_EXP = register(JobsPlus.getId("job_exp"));
+    IActionDataType<Double> JOB_EXP = register(JobsPlus.getId("job_exp"));
     IActionDataType<Job> ONLY_FOR_JOB = register(JobsPlus.getId("only_for_job"));
 
     static <T> IActionDataType<T> register(ResourceLocation location) {
